fixed compiling warning on VC

+++ lib_acl/include/stdlib/acl_define_win32.h
@@ -145,7 +145,7 @@ typedef int socklen_t;
 # define	ACL_DLL_FARPROC		FARPROC
 # ifndef	HAS_SSIZE_T
 #  define	HAS_SSIZE_T
-typedef long ssize_t;
+typedef intptr_t ssize_t;
 # endif
 
 # define ACL_INTERNAL_LOCK	ACL_FLOCK_STYLE_FLOCK
